fix: move CardSection component to CardSection folder
